requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.25.2
aiohttp==3.8.6
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.30.0
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bidict==0.23.1
boto3==1.37.31
botocore==1.37.31
CacheControl==0.12.14
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==1.9.2
cryptography==44.0.2
cycler==0.12.1
dask==2022.10.2
# dateparser==1.2.1 # Removed due to pytz conflict with rasa 3.5.14
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.11.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.7
libclang==18.1.1
locket==1.0.0
magic-filter==1.0.12
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.3
mattermostwrapper==2.2
msgpack==1.1.0
multidict==5.2.0
networkx==2.6.3
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
portalocker==2.10.1
prompt-toolkit==3.0.28
propcache==0.2.0
protobuf==3.19.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.2
pydot==1.4.2
PyJWT==2.9.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.1.4
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.5.14
rasa-sdk==3.5.1
redis==4.6.0
regex==2022.10.31
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
s3transfer==0.11.4
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
sentry-sdk==1.14.0
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
slack_sdk==3.35.0
SQLAlchemy==1.4.54
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.11.0
termcolor==2.4.0
tqdm==4.66.4
Twilio==7.14.2
typeguard==2.13.3
typing_extensions==4.13.0
tzdata==2025.1
tzlocal==5.2
uc-micro-py==1.0.3
urllib3==1.26.20
uvloop==0.17.0
websockets==10.4
Werkzeug==2.2.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
